// apps/cms/src/shared/api/generated/members/schemas/accountDirectoryItemResponseAccountType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea CMS Backend API — Members subset
 * Filtered for CMS member management Orval codegen.
 * OpenAPI spec version: v9
 */

/**
 * 계정 도메인
 */
export type AccountDirectoryItemResponseAccountType = typeof AccountDirectoryItemResponseAccountType[keyof typeof AccountDirectoryItemResponseAccountType];


export const AccountDirectoryItemResponseAccountType = {
  MEMBER: 'MEMBER',
  ADMIN_ACCOUNT: 'ADMIN_ACCOUNT',
} as const;